change add,delete favorite into toggle favorite

+++ src/controller/favorite.controller.js
@@ -1,22 +1,6 @@
 import {Favorite} from "../models/favorite.model.js";
 import { Song } from "../models/song.model.js";
 
-export const addFavoriteSong = async (req, res, next) => {
-    
-  try {
-    const { user, song } = req.body;
-    const exists = await Favorite.findOne({ user, song });
-    if (exists) return res.status(400).json({ error: 'Already favorited' });
-
-    const favorite = new Favorite({ user, song });
-    await favorite.save();
-    res.status(201).json(favorite);
-  } catch (error) {
-    next(error);
-  }
-
-}
-
 export const getFavoriteSongs = async (req, res, next) => {
     try {
         const {user} = req.params;
@@ -24,25 +8,29 @@ export const getFavoriteSongs = async (req, res, next) => {
         const favorites = await Favorite.find({ user: user }).populate("song");
         const favoriteSongs = await Song.find({ _id: { $in: favorites.map(fav => fav.song) } }).populate("artist").populate("album");
 
-        res.status(200).json({favorites, favoriteSongs});
+        res.status(200).json({favoriteSongs});
     } catch (error) {
         next(error);
     }
 }
 
-export const removeFavoriteSong = async (req, res, next) => {
-    
-try {
-    const { user, song } = req.params;
-    const deletedFavorite = await Favorite.findOneAndDelete({ user, song });
+export const toggleFavoriteSong = async (req, res, next) => {
+  try {
+    const { user, song } = req.body;
 
-    if (!deletedFavorite) {
-        return res.status(404).json({ message: "Favorite not found" }); // Handle case where no record is found
+    const exists = await Favorite.findOne({ user, song });
+
+    if (exists) {
+      // Remove from favorites
+      await Favorite.findOneAndDelete({ user, song });
+      return res.status(200).json({ message: "Song removed from favorites", favorited: false });
+    } else {
+      // Add to favorites
+      const favorite = new Favorite({ user, song });
+      await favorite.save();
+      return res.status(201).json({ message: "Song added to favorites", favorited: true });
     }
-    
-    res.status(200).json({message: "Song removed from favorites"});
-  } catch (error) {
-    next(error);
-  }
-
-}
+  } catch (error) {
+    next(error);
+  }
+};
